+++ src/syphon.inputreaders.js
@@ -3,20 +3,20 @@
 
 // Input Readers are used to extract the value from
 // an input element, for the serialized object result
-Backbone.Syphon.InputReaderSet = Backbone.Syphon.TypeRegistry.extend();
+Syphon.InputReaderSet = Syphon.TypeRegistry.extend();
 
 // Built-in Input Readers
-Backbone.Syphon.InputReaders = new Backbone.Syphon.InputReaderSet();
+Syphon.InputReaders = new Syphon.InputReaderSet();
 
 // The default input reader, which uses an input
 // element's "value"
-Backbone.Syphon.InputReaders.registerDefault(function($el){
-  return $el.val();
+Syphon.InputReaders.registerDefault(function(el){
+  return $(el).val();
 });
 
 // Checkbox reader, returning a boolean value for
 // whether or not the checkbox is checked.
-Backbone.Syphon.InputReaders.register("checkbox", function($el){
-  var checked = $el.prop("checked");
+Syphon.InputReaders.register("checkbox", function(el){
+  var checked = el.checked;
   return checked;
 });