+++ src/syphon.keyassignmentvalidators.js
@@ -6,16 +6,16 @@
 // extracted from the element. This is the last opportunity to prevent
 // bad data from getting serialized to your object.
 
-Backbone.Syphon.KeyAssignmentValidatorSet = Backbone.Syphon.TypeRegistry.extend();
+Syphon.KeyAssignmentValidatorSet = Syphon.TypeRegistry.extend();
 
 // Build-in Key Assignment Validators
-Backbone.Syphon.KeyAssignmentValidators = new Backbone.Syphon.KeyAssignmentValidatorSet();
+Syphon.KeyAssignmentValidators = new Syphon.KeyAssignmentValidatorSet();
 
 // Everything is valid by default
-Backbone.Syphon.KeyAssignmentValidators.registerDefault(function(){ return true; });
+Syphon.KeyAssignmentValidators.registerDefault(function(){ return true; });
 
 // But only the "checked" radio button for a given
 // radio button group is valid
-Backbone.Syphon.KeyAssignmentValidators.register("radio", function($el, key, value){ 
-  return $el.prop("checked");
+Syphon.KeyAssignmentValidators.register("radio", function(el, key, value){
+  return el.checked;
 });